backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
deep-translator==1.11.4
aiohttp==3.12.15
psutil==5.9.6
openai>=1.0.0
SQLAlchemy==2.0.36
selenium==4.15.2
webdriver-manager==4.0.1
PyJWT==2.8.0
bcrypt==4.1.2